refactor create conf package
