+++ pkg/conf/conf.go
@@ -1,9 +1,10 @@
-package aliases
+package conf
 
 import (
 	"errors"
 	"fmt"
 	"github.com/google/uuid"
+	"github.com/k-kinzal/aliases/pkg"
 	"github.com/k-kinzal/aliases/pkg/context"
 	"github.com/k-kinzal/aliases/pkg/docker"
 	"github.com/k-kinzal/aliases/pkg/util"
@@ -35,7 +36,7 @@ func LoadConfFile(ctx *context.Context) (*AliasesConf, error) {
 		return nil, fmt.Errorf("configuration file cannot read `%q`", err)
 	}
 
-	defs, err:= UnmarshalConfFile(buf)
+	defs, err:= aliases.UnmarshalConfFile(buf)
 	if err != nil {
 		return nil, err
 	}
