+++ pkg/export/alias.go
@@ -2,7 +2,7 @@ package export
 
 import (
 	"fmt"
-	"github.com/k-kinzal/aliases/pkg"
+	"github.com/k-kinzal/aliases/pkg/conf"
 	"github.com/k-kinzal/aliases/pkg/context"
 	"github.com/k-kinzal/aliases/pkg/docker"
 	"os"
@@ -11,7 +11,7 @@ import (
 )
 
 
-func Aliases(ctx *context.Context, conf *aliases.AliasesConf) {
+func Aliases(ctx *context.Context, conf *conf.AliasesConf) {
 	dir := ctx.GetBinaryPath()
 	os.Remove(dir)
 	os.Mkdir(dir, 0755)
