+++ pkg/export/path.go
@@ -2,12 +2,12 @@ package export
 
 import (
 	"fmt"
-	"github.com/k-kinzal/aliases/pkg"
+	"github.com/k-kinzal/aliases/pkg/conf"
 	"github.com/k-kinzal/aliases/pkg/context"
 	"os"
 )
 
-func Path(ctx *context.Context, conf *aliases.AliasesConf) {
+func Path(ctx *context.Context, conf *conf.AliasesConf) {
 	dir := ctx.GetBinaryPath()
 	os.Remove(dir)
 	os.Mkdir(dir, 0755)
